requirements.txt
aiogram>=3.0.0
sqlmodel>=0.0.14
psycopg[binary]>=3.1.0
sentry-sdk>=1.32.0
python-dotenv>=1.0.0
alembic>=1.12.0
sqlalchemy[asyncio]>=2.0.0
openai>=1.0.0
aiohttp>=3.8.0
